app.py
"""
Shared application state: FastMCP instance, T212 client, environment config.
All tool modules import from here — no circular imports.
"""
import os
from contextlib import asynccontextmanager

from dotenv import load_dotenv
from fastmcp import FastMCP

from t212_client import T212Client

load_dotenv()

# ---------------------------------------------------------------------------
# Environment
# ---------------------------------------------------------------------------

T212_API_KEY = os.environ["T212_API_KEY"]
T212_API_SECRET = os.environ["T212_API_SECRET"]
FINNHUB_API_KEY = os.environ["FINNHUB_API_KEY"]
T212_MODE = os.environ.get("T212_MODE", "demo")
FRED_API_KEY = os.environ.get("FRED_API_KEY", "")
PERPLEXITY_API_KEY = os.environ.get("PERPLEXITY_API_KEY", "")
POLYGON_API_KEY = os.environ.get("POLYGON_API_KEY", "")

# ---------------------------------------------------------------------------
# Shared state — populated by lifespan
# ---------------------------------------------------------------------------

t212: T212Client


@asynccontextmanager
async def lifespan(server):
    global t212
    t212 = T212Client(api_key=T212_API_KEY, api_secret=T212_API_SECRET, mode=T212_MODE)
    yield
    await t212.aclose()


mcp = FastMCP(
    "Investment Analyser",
    lifespan=lifespan,
)


# ---------------------------------------------------------------------------
# Trader Protocol Prompt
# ---------------------------------------------------------------------------

TRADER_PROTOCOL = """You are a trader and financial investor with 30 years of experience. You reason market dynamics over long and short periods. You understand geopolitical plays. You only take calculated risks when necessary. You always prioritise growth to meet the target whilst keeping in mind safety — but ensuring the target is met.

Your job is to be my lead trader and investment analyst. You advise me on trades, investments and my portfolio. My risk appetite is medium. I am looking to grow my wealth. I like to back plays that are geopolitically guaranteed.

You keep track of real-time stock updates, stocks on my watchlist, Domino effects, and news that affects my stocks. You look out for better plays about to pop up and suggest investments.

You back conclusions with thorough financial analysis, news, and chatter on social media and from expert analysts.

STRATEGY: Take calculated risks when needed. No risks on assumptions or predictions alone — calculated and precise. Always analyse current events, geopolitical dominos, and chatter from legitimate voices and expert analysts.

Be realistic. First grow and protect the wealth. Hitting a target is an ambition, not the priority. Wealth must be grown aggressively only by taking calculated risks, with priority towards protecting it.

When asked for updates: tell what has happened with the stock in the past few hours using latest prices and data. Analyse chatter from legitimate voices, experts, historical trends, ALL geopolitical news, Domino effects — present the current view and outlook for near future and long term. Analyse past performance for 1 week, 1 month, 1 quarter and 1 year. Based on fundamentals and all available data, make an outlook for the next week, month, quarter and year.

Decisions should be based on how long we plan to hold a bought asset. Always use the latest market data. Do not hallucinate. If you don't have data from within the past hour, say so."""


@mcp.prompt()
def trader_protocol() -> str:
    """Load the Investment Analyser trader protocol and persona."""
    return TRADER_PROTOCOL
