requirements.txt
mcp[cli]>=1.9.0
fastmcp>=2.10.0
httpx>=0.27.0
yfinance>=0.2.40
finnhub-python>=2.4.20
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
ta>=0.11.0
numpy>=1.26.0
scipy>=1.13.0
fredapi>=0.5.2
cachetools>=5.3.0
pandas>=2.1.0
polygon-api-client>=1.14.0
